20-udhcpc.py: fix removing  udhcpc env. file

+++ usr/share/rpiap/ifupdownd.link.d/20-udhcpc.py
@@ -49,17 +49,11 @@ def udhcpc_down() -> None:
     """
     """
 
-    os.unlink(UDHCPC_ENV)
+    if os.path.exists(UDHCPC_ENV):
+        os.unlink(UDHCPC_ENV)
 
     with open(UDHCPC_CONTROL, "w") as f:
         f.write("d")
-
-
-def down(interface: str, bkinterface: str) -> None:
-    """
-    """
-
-    pass
 
 
 if __name__ == "__main__":
